requirements.txt
streamlit
llama-index
llama-index-llms-ollama
llama-parse
pydantic
python-dotenv
llama-index-tools-google
google-auth-oauthlib
google-api-python-client
llama-index-embeddings-ollama
pyyaml